docker-captcha/requirements.txt
Flask==3.0.0
pycryptodome==3.20.0
Crypto==1.4.0
PyExecJS
opencv-python
